Separate chat messages into their own resource controller
