+++ app/Http/Controllers/Chat/ChannelsController.php
@@ -20,14 +20,10 @@
 
 namespace App\Http\Controllers\Chat;
 
-use App\Exceptions\API;
 use App\Models\Chat\Channel;
-use App\Models\Chat\Message;
 use App\Models\Chat\UserChannel;
-use App\Models\User;
 use Auth;
 use DB;
-use Request;
 
 class ChannelsController extends Controller
 {
@@ -38,45 +34,6 @@ class ChannelsController extends Controller
             'Chat/Channel'
         );
     }
-
-    public function show($channel_id)
-    {
-        $user_id = Auth::user()->user_id;
-        $since = Request::input('since');
-        $limit = clamp(Request::input('limit', 50), 1, 50);
-
-        $userChannel = UserChannel::where(['user_id' => $user_id, 'channel_id' => $channel_id])->firstOrFail();
-
-        if ($userChannel->channel->isPM()) {
-            // restricted users should be treated as if they do not exist
-            if (optional($userChannel->channel->pmTargetFor(Auth::user()))->isRestricted()) {
-                abort(404);
-            }
-        }
-
-        $messages = $userChannel->channel
-            ->filteredMessages()
-            ->with('sender');
-
-        if (presence($since)) {
-            $messages = $messages->where('message_id', '>', $since)
-                ->orderBy('message_id', 'asc')
-                ->limit($limit)
-                ->get();
-        } else {
-            $messages = $messages->orderBy('message_id', 'desc')
-                ->limit($limit)
-                ->get()
-                ->reverse();
-        }
-
-        return json_collection(
-            $messages,
-            'Chat\Message',
-            ['sender']
-        );
-    }
-
     public function join($channel_id, $user_id)
     {
         // FIXME: Update this to proper permission check when public-only restriction is lifted
@@ -117,37 +74,5 @@ class ChannelsController extends Controller
         $userChannelQuery->update(['last_read_id' => DB::raw("GREATEST(COALESCE(last_read_id, 0), $message_id)")]);
 
         return response([], 204);
-    }
-
-    public function send($channel_id)
-    {
-        $channel = Channel::findOrFail($channel_id);
-
-        if ($channel->isPM()) {
-            // restricted users should be treated as if they do not exist
-            if (optional($channel->pmTargetFor(Auth::user()))->isRestricted()) {
-                abort(404);
-            }
-        }
-
-        priv_check('ChatChannelSend', $channel)->ensureCan();
-
-        try {
-            $message = $channel->receiveMessage(
-                Auth::user(),
-                Request::input('message'),
-                get_bool(Request::input('is_action', false))
-            );
-        } catch (API\ChatMessageTooLongException $e) {
-            abort(422, $e->getMessage());
-        } catch (API\ExcessiveChatMessagesException $e) {
-            abort(429, $e->getMessage());
-        }
-
-        return json_item(
-            $message,
-            'Chat/Message',
-            ['sender']
-        );
     }
 }
